코드 리뷰에 따라서 todos 모듈화 및 전역변수 제거
1. todos 객체를 todos.js파일로 모듈화하여 사용
2. 전역객체 todos를 직접 참조하던 함수들을 인자를 받아 사용하는 방식으로 수정
3. checkTags()의 foreach문을 filter와 map을 사용하는 것으로 수정
